requirements.txt
boto3==1.9.15
botocore==1.12.15
dateparser==0.7.0
docutils==0.14
jmespath==0.9.3
numpy==1.15.2
pandas==0.23.4
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
regex==2018.8.29
s3fs==0.1.6
s3transfer==0.1.13
six==1.11.0
tzlocal==1.5.1
urllib3==1.23
